backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
httpx==0.25.2
osmnx==1.6.0
python-pptx==0.6.23
playwright==1.40.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
folium==0.15.0
requests==2.31.0
geojson-pydantic==0.6.0
